save soll - ist done

+++ database/migrations/2023_04_28_092759_create_course_subject_pivot_table.php
@@ -13,7 +13,7 @@ class CreateCourseSubjectPivotTable extends Migration
             $table->unsignedBigInteger('subject_id');
             $table->unsignedBigInteger('template_id'); // include template_id
             $table->integer('soll')->nullable();
-            $table->integer('ist')->default(0);
+            $table->float('ist')->default(0);
             $table->timestamps();
 
             // Foreign keys
